Complete the paulisum move
