+++ crates/ppvm-paulisum/Cargo.toml
@@ -4,7 +4,6 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-ahash = { version = "0.8.12", optional = true }
 approx = { version = "0.5.1", optional = true }
 bitvec = "1.0.1"
 bon = "3.7.2"
@@ -23,11 +22,11 @@ insta = { version = "1.43.2", features = ["yaml"] }
 
 [features]
 default = ["dashmap", "gxhash", "ahash", "indexmap", "approx"]
-ahash = ["dep:ahash"]
+ahash = ["ppvm-runtime/ahash"]
 approx = ["dep:approx"]
-dashmap = ["dep:dashmap", "dep:rayon"]
-gxhash = ["dep:gxhash"]
-indexmap = ["dep:indexmap"]
+dashmap = ["dep:dashmap", "dep:rayon", "ppvm-runtime/dashmap"]
+gxhash = ["dep:gxhash", "ppvm-runtime/gxhash"]
+indexmap = ["dep:indexmap", "ppvm-runtime/indexmap"]
 
 [[bench]]
 name = "gate"
