+++ crates/ppvm-paulisum/tests/ghz.rs
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2026 The PPVM Authors
 // SPDX-License-Identifier: Apache-2.0
 
-use ppvm_runtime::prelude::*;
+use ppvm_paulisum::prelude::*;
 
 #[test]
 fn test_h() {
