+++ crates/ppvm-paulisum/tests/preserve.rs
@@ -9,9 +9,9 @@ use std::collections::HashSet;
 
 use ppvm_runtime::config::Config;
 use ppvm_runtime::config::fxhash::ByteF64;
-use ppvm_runtime::prelude::*;
-use ppvm_runtime::strategy::CoefficientThreshold;
-use ppvm_runtime::sum::PauliSum;
+use ppvm_paulisum::prelude::*;
+use ppvm_paulisum::strategy::CoefficientThreshold;
+use ppvm_paulisum::sum::PauliSum;
 
 type Cfg = ByteF64<1>;
 type CfgThr = ByteF64<1, CoefficientThreshold>;
