+++ crates/ppvm-paulisum/tests/strategy.rs
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2026 The PPVM Authors
 // SPDX-License-Identifier: Apache-2.0
 
-use ppvm_runtime::{
+use ppvm_paulisum::{
     prelude::*,
     strategy::{CoefficientThreshold, CombinedStrategy, MaxPauliWeight},
 };
