+++ crates/ppvm-runtime/Cargo.toml
@@ -4,12 +4,17 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+ahash = { version = "0.8.12", optional = true }
 anyhow = "1.0.100"
 bincode = { version = "2.0.1", optional = true }
 bitvec = "1.0.1"
+dashmap = { version = "6.1.0", features = ["rayon"], optional = true }
 fxhash = "0.2.1"
+gxhash = { version = "3.5.0", features = ["deterministic"], optional = true }
+indexmap = { version = "2.11.4", optional = true }
 itertools = "0.14.0"
 num = "0.4.3"
+rayon = { version = "1.11.0", optional = true }
 serde = { version = "1.0.228", optional = true }
 
 [dev-dependencies]
@@ -17,5 +22,9 @@ insta = { version = "1.43.2", features = ["yaml"] }
 
 [features]
 default = []
+ahash = ["dep:ahash"]
 bincode = ["dep:bincode"]
+dashmap = ["dep:dashmap", "dep:rayon"]
+gxhash = ["dep:gxhash"]
+indexmap = ["dep:indexmap"]
 serde = ["dep:serde"]
