+++ crates/ppvm-runtime/src/map/mod.rs
@@ -2,3 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 mod hashmap;
+
+#[cfg(feature = "dashmap")]
+mod dashmap;
