+++ crates/ppvm-sym/Cargo.toml
@@ -7,4 +7,5 @@ edition = "2024"
 anyhow = "1.0.100"
 fxhash = "0.2.1"
 num = "0.4.3"
+ppvm-paulisum = { version = "0.1.0", path = "../ppvm-paulisum", default-features = false }
 ppvm-runtime = { version = "0.1.0", path = "../ppvm-runtime", default-features = false }
